src/test/java/com/mitrais/cdc/bloggatewayapplicationv1/services/AuthenticationServiceTest.java
package com.mitrais.cdc.bloggatewayapplicationv1.services;

import com.mitrais.cdc.bloggatewayapplicationv1.payload.AuthenticationPayload;
import com.mitrais.cdc.bloggatewayapplicationv1.payload.LoginResponse;
import com.mitrais.cdc.bloggatewayapplicationv1.payload.TokenPayload;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

@RunWith(SpringRunner.class)
@SpringBootTest
public class AuthenticationServiceTest {

    @Autowired
    AuthenticationService authenticationService;

    @Test
    public void login() {
        AuthenticationPayload authenticationPayload = new AuthenticationPayload("admin", "admin123");
        LoginResponse loginResponse =  authenticationService.login(authenticationPayload);

        assertThat("admin", is(loginResponse.getData().getUsername()));
        assertThat("You have login successfully", is(loginResponse.getMessage()));

    }
}